Detailed env suggestions for l2
- Using `l2 --env ${l2FilePath}` go lang binary command to fetch env.
- Removed old searching envs in the frontend as performance was low.
- Moved all the suggestions and replacing part to new file.
